+++ helpers/IDProvider.mjs
@@ -1,0 +1,5 @@
+export default class IDProvider {
+    static random() {
+        return Math.random().toString(16).slice(2);
+    }
+}