+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 
+## [26.3.0] - 2026-04-02
+
+### Added
+
+- Style registry
+- Random ID provider
+- Fontawesome icons component
+- Core components
+- Button component
+- Buttons group component
+
+### Breaking
+
+- Reworked entire system. Components will come back little by little
+
+
+
 ## [25.2.4] - 2025-12-11
 
 ### Added